+++ app/src/main/java/com/zxm/okhttp/entiy/BaseUserInfo.java
@@ -1,4 +1,4 @@
-package com.zxm.okhttp;
+package com.zxm.okhttp.entiy;
 
 import lombok.Data;
