+++ app/src/main/java/com/zxm/okhttp/entiy/UserInfo.java
@@ -1,4 +1,4 @@
-package com.zxm.okhttp;
+package com.zxm.okhttp.entiy;
 
 import lombok.Data;
 
@@ -6,10 +6,6 @@ import lombok.Data;
 public class UserInfo {
     private String userId;
     private String userName;
-    private String nickName;
-    private String role;
     private String email;
     private String token;
-    private String createdAt;
-    private String updatedAt;
 }
