+++ app/src/main/java/com/zxm/okhttp/net/ApiErrorModel.java
@@ -1,4 +1,4 @@
-package com.zxm.okhttp;
+package com.zxm.okhttp.net;
 
 import lombok.Data;
